feat(z-range-picker): move functions to utils.tsx
